refactor(bench): dlopen-based isolation; switch to canonical scheme repos
Each parameter set is now a self-contained .so loaded with
dlopen(RTLD_LOCAL). Non-namespaced symbol conflicts are impossible
regardless of the underlying implementation's naming conventions.

Contract every .so must satisfy:
  const bench_scheme_info_t *bench_info(void);  // metadata
  int crypto_sign_keypair(uint8_t *pk, uint8_t *sk);
  int crypto_sign_signature(...);               // standard NIST signature
  int crypto_sign_verify(...);

New files:
  loader.h / loader.c   — dlopen + dlsym; populates bench_scheme_t
  scheme.h              — bench_scheme_info_t (used by shims, not harness)
  harness.h             — unchanged interface; now includes loader.h

ML-DSA: replaced monolithic mldsa.c+h with three shims
  (mldsa44/65/87_shim.c). Each shim forward-declares dilithium's
  namespaced API and re-exports under crypto_sign_* names.
  Source: pq-crystals/dilithium (existing submodule, unchanged).

SLH-DSA: removed PQClean-based implementation. Submodule URL
  registered (pq-code-package/slhdsa-c); Makefile scaffolded with
  check-submodule guard. Shims to be added once network available.

main.c: now calls bench_load/bench_unload per .so path; adding a
  scheme requires only building its .so and appending to SO_PATHS[].

+++ bench/harness.h
@@ -1,11 +1,9 @@
 #pragma once
 #include <inttypes.h>
-#include <stddef.h>
 #include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <string.h>
-#include "scheme.h"
+#include "loader.h"
 
 /* ---------- cycle counter ---------- */
 
@@ -39,8 +37,8 @@ static int bench_cmp_u64(const void *a, const void *b) {
     return (x > y) - (x < y);
 }
 
-static uint64_t bench_median(uint64_t *arr, size_t n) {
-    qsort(arr, n, sizeof(uint64_t), bench_cmp_u64);
+static uint64_t bench_median(uint64_t *arr, int n) {
+    qsort(arr, (size_t)n, sizeof(uint64_t), bench_cmp_u64);
     return (n & 1) ? arr[n / 2] : (arr[n / 2 - 1] + arr[n / 2]) / 2;
 }
 
@@ -58,9 +56,9 @@ static void bench_run(const bench_scheme_t *s) {
     uint8_t  msg[32] = {0};
     size_t   siglen  = 0;
 
-    uint64_t *kg_cyc = malloc(n * sizeof(uint64_t));
-    uint64_t *sg_cyc = malloc(n * sizeof(uint64_t));
-    uint64_t *vf_cyc = malloc(n * sizeof(uint64_t));
+    uint64_t *kg_cyc = malloc((size_t)n * sizeof(uint64_t));
+    uint64_t *sg_cyc = malloc((size_t)n * sizeof(uint64_t));
+    uint64_t *vf_cyc = malloc((size_t)n * sizeof(uint64_t));
 
     if (!pk || !sk || !sig || !kg_cyc || !sg_cyc || !vf_cyc) {
         fprintf(stderr, "bench_run: allocation failed for %s\n", s->name);
